refactor graphql code into github helper class
